refactor(benchmark): organize the drafts page - Keep the page entry in index.tsx - Place draft rows in the page components directory
